feat: unconstrained variants for event emission (#7251)
This pr adds an `unconstrained` function for encoding and encrypting
note/event logs.

Can be used when you don't need a strong guarantee of delivery, but
trust the sender. Could for example be if I am buying a table from you.
If you receive the funds, and can see them, then I encrypted the correct
thing, and you will hand over the table afterwards. But should not be
used when using a contract for doing an atomic swap etc, where you could
end up with bad data for the lulz.