chore: included subrelation witness degrees in the relations relevant to zk-sumcheck (#7479)
The subrelation witness degrees are used in zk-sumcheck and affect the
sizes of the univariates in the SumcheckTupleOfTuplesOfUnivariates as
well as the sizes of the sumcheck round univariates.

In several relations, I used upper bounds on the witness degrees which
are simply given by the corresponding subrelation's length - 1.
Indicated explicitly that those are upper bounds. Computed the precise
constants for the relations relevant to Mega.

The auxiliary and logderiv_lookup_relation contain a trick allowing to
re-use the values accumulated by an accumulator of a different size
using ShortView.