refactor(bb): simplify parallel_for_if_effective (#8079)
The chunkiness of `run_loop_in_parallel_if_effective` was getting in the
way of doing polynomial refactor passes for structured polynomials.

- Renamed parallel for variants to all start with parallel_for
- Doing math on the spot with constants was simpler, plus other
simplifications.
- Removed the variant that takes two parameters, and introduced a one
parameter variant as just having an 'i' parameter felt like the sweet
spot, plus a lot of places where the thread/chunk index was not used I
do plan to use in a followup as we can avoid mutexes that way and be
more efficient.

Bundled changes:
- There was an unnecessary polynomial copy in IPA
- Also, introduce a SlabVector class where we were using
SlabContainerAllocator, with plans to use it more (and possibly make it
equal to normal std::vector for native code where memory fragmentation
really doesn't matter)