refactor: Renaming around Protogalaxy Prover (#8272)
This is literally just a ton of renaming + shuffling some declaration
and defs in the Protogalaxy prover so the orders of these match. Some
highlights:
- I wanted to stop using the term "optimised" around the Pg optimization
that skips computing zero because we have many optimizations and the
name was unclear. I also put the extra qualifier on the non-production
case where we _don't_ use that optimization, which is currently just in
tests of the combiner.
- Ariel told me that the real name of the protocol is Protogalaxy--this
is in the name of the paper in pdf form, but not on the eprint page
:shrug:
- `PowPolynomial` is now a misnomer because it doesn't involve powers of
a challenge $\beta$, but rather a bunch of monomials generate from a set
of $\beta_i$'s. So I rename this and the corresponding files. I didn't
have to change this but I do think clarity here comes at a low enough
cost to merit making the change.
- We have functions that refer to the "full honk" relation, which is
fine, but strictly speaking there is no connection to Honk when using
Protogalaxy or sumcheck in insolation, so it makes sense to give more
agnostic and IMO slightly clearer names.