chore: GitHub Actions Deployments to Amazon EKS (#8563)
This GH Actions yaml files to allow network pushes to the new Spartan
Kubernetes cluster.
- This action is triggered when branches are merged into `staging` and
`production` branches of the repo.
- Intended for maintenance of permanent network deployments. `Helm`
either installs or updates, but does not delete the long running
`staging` and `production` environments.
- Manual triggering is intentionally disabled to prevent
partial/corrupted/conflicting `helm install` instances (helm does not
support a lockfile or otherwise protect against this natively).
- The current `helm install` command does not set any values. This can
be changed in the future to support `staging` and `production`
configurations that differ from what runs on local and ci test
environments.