feat: more robust recursion input generator (#8634)
Introduces the method `write_recursion_inputs_honk` in main.cpp that,
given a program, produces the inputs (in the form of a Prover.toml) to a
second noir program that recursively verifies a proof of the first. This
is used to update/simplify the logic in
`regenerate_verify_honk_proof_inputs.sh` as a proof of concept. (Also
replaces `update_verify_honk_proof_inputs.py` with similar logic in
`ProofSurgeon` that deals more dynamically with the public inputs).

The reason for doing this is that similar logic will be needed to create
tests for `noir::verify_proof()` for Oink/PG which will have slightly
different forms. A flow of this kind will also be used to generate
inputs from typescript in the integration and e2e tests.