feat(avm)!: GETENVVAR + ISSTATICCALL (#8692)
The reason I'm keeping all the oracles on the Noir side, is because I
need to pass the enum index as an immediate and Noir doesn't allow that
in any other way. That is, if you had an oracle `get_env_var(var_idx:
u8)`, when you call it like `get_env_var(12)` you would get a memory
address with the 12 in it; and not the `12` constant.