feat!: remove SharedMutablePrivateGetter (#8749)
Missed this deletion while removing the key registry. I slightly
adjusted one of the e2e tests that used to check certain private getter
properties so that it still performs the same checks on the regular
shared mutable, as that doesn't seem to be present in any other e2e
test.