fix: Spartan kubernetes cluster IaC (#8893)
This PR includes the Terraform IaC used to create the current Spartan
AWS cluster and fixes a previous error where server nodes were being
repeatedly terminated and re-added to the cluster. This code can be used
to make future modifications to the cluster, and the backend state file
has been saved to S3 for persistence and sharing between engineers.

`variable.tf` and `outputs.tf` files have been added for better
organization in the event the cluster configuration grows larger.