feat: Replace Zeromorph with Shplemini in ECCVM (#9102)
This PR switches ECCVM to Shplemini which shaves off ~300k in the tube
circuit. Now, on the verifier side, we first execute Shplemini, then
reduce the BatchOpeningClaim to a single OpeningClaim by performing the
batch_mul delayed by Shplemini. Then, we construct the translation
OpeningClaim, and the two are being reduced to a single OpeningClaim by
executing a second iteration of Shplonk. Finally, we verify the
OpeningClaim via PCS. This could be further optimised as we currently
perform 4 batch_muls.