feat: modify contract instance to include public keys (#9153)
In this PR we are doing the ground work for the new address scheme by
modifying the contract instance to include the full public keys instead
of only the public keys hash. We need the full public keys because we
need to verify the preimage of the new address, which requires the ivpk,
and we need to verify the ivpk's correctness by manually computing the
public keys hash.