feat(avm)!: returndatasize + returndatacopy (#9475)
This PR
* Introduces RETURNDATASIZE and RETURNDATACOPY (also copies revert data if reverted)
* Fixes a bug in CALL in witgen
* Changes the public context to return slices when calling public functions. This was partly done because templated functions would always be inlined and I wanted to avoid that blowup

Note that the rethrowing hack is still present in the simulator, so the rethrowing branch in the public context is still not used. I will make this change once @sirasistant finishes the string encoding changes.

In a later PR we can remove the returndata from CALL.

Part of #9061.