feat: change definition of lagrange last (#9916)
Update the `lagrange_last` polynomial to take 1 at the idx of the last
active wire in the execution trace. This effectively translates the
existing checks in the permutation relation to this index rather than
the dyadic size and allows the permutation grand product to be computed
only up to this point. This is a precursor for kernel VKs that are
independent of the ambient trace size.

---------

Co-authored-by: maramihali <mara@aztecprotocol.com>