test(avm): minor benchmarking (#9869)
Just a very small micro-benchmark for the a future PR to reference.

This sets up the smallest "worst-case" scenario for the current ephmeral
tree implementation.

1) An initial public data tree is seeded with 128 dummy leaves,
occupying slots 0 to 127
2) We update slot 0 with a new value - this causes the ephemeral tree to
track slot 0 as the `indexedTreeMin`.
3) Inserting 64 new slots (as per a tx), results in 126 DB accesses and
a linear (in the leaf count) search complexity
a) The DB reads occur because the ephemeral tree doesnt track slots 1 to
127
b) All reads are O(n) as we traverse from min leaf to the low leaf info