chore: Split up protocol contract artifacts (#10765)
The protocol contracts package was loading all artifacts in a single
file, and then re-exporting from there. This means that whenever a
protocol contract was loaded, all other protocol contracts were fetched
and processed as well.

This breaks protocol contracts into granular exports, and adds a bundle
export that contains all for the few cases where we do need a single
function for loading all artifacts.