fix: Block building test timeout (#10812)
The test `can simulate public txs while building a block` would, in some
cases, build blocks with varying tx sizes. An example run built blocks
with 4, 12, and 12 txs. This totals 28 txs, so the remaining 2 txs never
get mined and the test timeout.

This fixes it by forcing the sequencer to build with just 4 txs
consistently, and making the total number of txs a multiple of it.