chore: Configurable parallelism in bootstrap (#10909)
Bootstrap scripts defined RAYON_NUM_THREADS and HARDWARE_CONCURRENCY to
16, and also ran parallel with 16 processes. This means that a bootstrap
run could take up to 256 cores.

This PR loads the values for those vars from env if set, so they can be
overridden, but defaults to the original values that work on
CI/mainframe.

Also honors the env vars in test scripts.