chore: Block building benchmark via github-action-benchmark (#11202)
Deletes old benchmarks, along with the benchmark-related scripts and
types.

Adds a single benchmark for block building, with a stubbed
`TelemetryClient` that collects all datapoints in memory, and then
flushes a set of specified metrics into the custom format expected by
[github-action-benchmark](https://github.com/benchmark-action/github-action-benchmark),
which we're currently using for bb.

Benchmarks get published to
https://aztecprotocol.github.io/aztec-packages/dev/e2e-bench/

Fixes #11154