fix: handle sequencer building block mid-synch (#11735)
This PR fixes a race condition in the sequencer that would happen in e2e
tests (or sandbox, basically anywhere we warp time because of auto mine)
where the sequencer would not handle a block being synchronized in the
middle of block production.

This PR merges getting the chain tip with checking if its components are
synchronized. This way it at least will fail at `canProposeAtTime` since
the provided archive might be old.

Thanks @spalladino and @LHerskind for help debugging this one :)