chore: op wires index from 0 (#11986)
The ecc op relation in Mega is designed to check that the ecc op data
has been duplicated across the wires and ecc_op_wires. Because the wires
need to be shifted, they contain a zero row, meaning the ecc op data is
stored starting from index 1. Prior to this PR, the op wires also stored
the data this way, even though they dont need to be shifted. Its more
convenient for the ecc op wires to index from zero. This is achieved by
simply utilizing the shifted wires in the relation so that we check
ecc_op_wire[i] = wire[i+1].

Note: also fixes an incorrect skip condition in the EccOpRelation that
was not caught prior due to an unrelated bug that will be fixed in
another PR.