chore: @aztec/simulator reorg (#12361)
As part of a larger PXE refactor, moved components around
`@aztec/simulator` to better identify responsibilities and which code
goes where.

* `public` now contains all public code (including AVM simulator), but
it's not 100% of the `server` bundle. Same thing for the newly created
`private` folder.
* `/client` bundle contains MOST of `private` code, except for the
native and "with blobs" versions of `acvm` simulator. It also has all
the common code
* `/server` bundle contains all of `public` code, plus the native and
"with blobs" versions of the `acvm` simulator. Like client, it has all
the common code.

---------

Co-authored-by: esau <152162806+sklppy88@users.noreply.github.com>