feat: tightly pack logs inside blobs (#11752)
This PR removes trailing zeroes from logs appended to the blobs to save
on field space.
e.g. before this PR, a public log would be appended like:
```rust
[1, 2, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
```
but now is appended as:
```rust
[1, 2, 3]
```
with a length prefix.
In ts, we add back the trailing zeroes when constructing tx effects from
the blob, so if they are required there should be no issue.
Also, in some logs there are valid zeroes inside the log. The method
does not remove these. e.g. a public log like:
```rust
[1, 0, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
```
is appended as:
```rust
[1, 0, 3]
```
then reconstructed to the original log once in ts.