feat: Combine group polynomials in translator by interleaving rather than concatenation (#12343)
To facilitate making the translator ZK, we change the strategy of
combining the range constraint polynomials in translator (defined over
the `mini_circuit_size`) by interleaving rather than concatenation. This
way, when blinding the polynomials, all the randomness will be located
at the end. This PR changes the PCS protocol to handle interleaved
polynomials and ensures the Translator polynomials are constructed by
interleaving (synchronously for now). The changes were integrated in the
ClaimBatcher and PolynomialBatcher

Summary of changes required in the Gemini protocol.
With interleaving, the full batched polynomial is defined as A(X) = F(X)
+ G(X)/X + P(X^s) where s is the size of the group of polynomials to be
interleaved, assumed even. Before, the concatenated component was
partially evaluated as part of A_0+ and A_0-. Now we construct A_0+ and
A_0- only from the batched shifted and batched unshifted components and
separate polynomials P+ and P- for the interleaved components and send
two new opening claims for the evaluation of P+ and P- at r^s: (P+, r^s,
P(r^s)) and (P-, r^s, P-(r^s). This way we can correctly reconstruct
A(r) = A_0+(r) + P+(r^s) and A(-r) = A_0-(-r) + P_(r^s). The protocol
enforces s has to be even so that (-r)^s = r^s