refactor: remove selector from public call request (#12828)
- Remove `function_selector` from `PublicCallRequest` as it is no longer
used.
- Remove `public_functions` from `ContractClass`.
- `ContractClass` checks that there's at most 1 public function in the
artifact.
- Rename `args` to `calldata` when it's public calldata that includes
the function selector.
- Use different generator indexes for calldata and regular args.

### aztec-nr
- Rename oracle calls `enqueue_public_function_call_internal ` to
`notify_enqueued_public_function_call `, and
`set_public_teardown_function_call_internal` to
`notify_set_public_teardown_function_call `, to be consistent with other
oracle calls.
- Storing data to execution cache will need to provide the hash in
addition to the preimage, making it possible to use it for different
types of data. Currently it's used for calldata and args.

### node
- Rename `getContractFunctionName` to `getDebugFunctionName`, as the
function name won't always be available. It's set by explicitly calling
`registerContractFunctionSignatures` for debugging purpose.
- Remove `ExecutionRequest[]` in `Tx` and replace it with `calldata[]`.
We now loop through the `publicCallRequests` in the public inputs and
pair each with an entry in `calldata[]`.
- `DataValidator` checks that calldata are provided for all the
non-empty public call requests, and the hash is correct.