chore: remove dummy ops in decider pk (#13049)
We used to not serialise and deserialise infinity points in the
transcript properly until
[this](https://github.com/AztecProtocol/aztec-packages/pull/7709). More
precisely, it was possible to get to
[this](https://github.com/AztecProtocol/aztec-packages/blob/mm/handle-merge/barretenberg/cpp/src/barretenberg/ecc/groups/element_impl.hpp#L68)
point in the projective to affine constructor (used when operating with
zero commitments) because the transcript was not setting the infinity
flag in situations when `z=0`. This was leading to intermmitent issues
that resulted in us adding dummy ecc ops [this
PR](https://github.com/AztecProtocol/aztec-packages/pull/5174/files). We
should now be able to remove these dummy ops.