chore: Add type annotation so the inferred type isn't Field (#14507)
In preparation for https://github.com/noir-lang/noir/pull/8635 where the
format of `Field` printed at `comptime` will change from numeric to
hexadecimal, to match the behaviour of the runtime. A slight
inconvenience is that `let x = 0;` seems to be inferred to be `Field` by
the compiler, which would show up as `0x00`, which in this case
disrupted the macro, as what has been rendered as `arg0` became
`arg0x00`.