fix: Fixes race condition in public processor (#14605)
This PR fixes the processing of public functions. Previously there was a
race condition with transactions that went beyond the deadline. They
continued to update the world state whilst the sequencer was finalising
the block. The resulting block would be invalidated by the updates.

Here we introduce the ability to `commitAllCheckpoints` and
`revertAllCheckpoints` in the world state. Once the processing of public
functions completed, potentially with a tx timing out, the sequencer
stops further access to the world state by the processor and calls
`revertAllCheckpoints`. This ensures the sequencer has a consistent
world state to finalise the block.

---------

Co-authored-by: just-mitch <68168980+just-mitch@users.noreply.github.com>