feat: offchain msg oracle and tx effect (#14771)
Fixes #14747

Second of a sequence of PRs implementing [out-of-band event
delivery](https://github.com/AztecProtocol/aztec-packages/issues/14698)
(first one was
https://github.com/AztecProtocol/aztec-packages/pull/14634).

In this PR I implemented the following flow:
1. Message gets emitted from contract via the `emit_offchain_message`
oracle,
2. then in TS the messages are stored inside the
`PrivateExecutionOracle`,
3. once the execution finishes we collect the messages from the oracle
and pass them to the constructor of `PrivateCallExecutionResult`,
4. then once the tx is proven a `TxProvingResult` is constructed and
`TxProvingResult` has a new getter that collects all the offchain
messages,
5. this getter is called once the tx is proven and the messages are then
stored inside `ProvenTx`.

**Note 1:** It didn't make sense to have the offchain messages be stored
on the Tx object as that is the thing that is transmitted to the p2p
network via aztec node.
**Note 2:** I have not yet figured out the `Aztec.js` API (i.e. how to
expose this from the `send(...)` method). That will be done in the
following PR. For now the `ProvenTx` was the logical last step to do in
this one (the messages "travelled" there "naturally").

---------

Co-authored-by: Nicolás Venturo <nicolas.venturo@gmail.com>