feat: batch partial note completion processing (#15051)
This makes partial notes be completed concurrently by creating a capsule
array of `LogRetrievalRequest`s, which is read and processed via
`Promise.all`, with the results later stored in a second capsule array
of `Option<LogRetrievalResponse>`s. This was a bit annoying since I had
to introduce these intermediate structs, serialization tests, etc., as
well as do some tidying up of missing bits like serialization of
`Option`s.

Overall, the thing is not _too_ bad. I deleted the old oracles since we
no longer need them given this construct, which also let me get rid of
the weird array re-sizing we had going on that dealt with varying log
lengths depending on source. It is a massive abuse of CapsuleArrays, and
another good example of the need for #14943, since we get responses that
we need to manually clear, etc.

Closes https://github.com/AztecProtocol/aztec-packages/issues/14710