feat: Tx collection service (#15259)
Adds a new `TxCollection` service that handles gathering txs via reqresp
or by hitting preconfigured nodes via RPC. Replaces the old
`TxCollector` for validators, and `ProverCoordination` for provers.

This service runs continuosly, keeping a backlog of txs to collect
(populated by all mined txs not found on the local tx pool), and
querying the nodes and reqresp in a "slow" loop. The service also
provides methods for "fast" time-bound collection, which is used for
validators to collect all txs for an attestation, or a prover to collect
all txs for proving. It also monitors events from the tx pool (now an
event emitter) to flag any new txs as found, in case they are received
via other means (eg gossip). It reports metrics on how many txs are
found via the fast or slow loops, via reqresp or nodes, and the duration
for each individual request.

Note that `TxCollection` only deals with fetching txs via reqresp or
nodes. It does not handle the tx pool. For that, a new `TxProvider` is
added, which extracts txs from the tx pool, from block proposals (when
querying txs for attestation), or from the tx collection service. As
opposed to `TxCollection`, the `TxProvider` is stateless. Both of these
components are currently managed by the `P2PClient`, though we should
extract them (along with the tx pool and tx lifecycle management) to a
separate subsystem (see #14269).

Given the `P2PClient` orchestrates tx collection, the `txFetcher` loop
from the prover node is now removed, and the prover node is set up such
that it always spins up a `P2PClient`, even if it is just for tx
collection. If p2p is disabled, the P2PClient is initialized with a
dummy p2p service.

As an unrelated change, this PR also adds a `TxWithHash` type, because I
hated having to `await` every single request to get a tx hash,
especially when I knew it was already precomputed. Eventually I'd want
to have the hash always precomputed on a `Tx`, so we don't have to await
`getHash`, and then we can delete this new type.

Fixes #15102
Fixes #15103
Fixes #15104
Fixes #15105

Pending tasks:
- Report txs as found as soon as requested from nodes
- Fix broken e2e tests
- Add tests for tx collection (using the ones from prover coordination
as inspiration)
- Remove prover coordination
- Fix conflicts