chore: deflake data withholding (#15537)
Fixes an issue in the data withholding test where the prune would happen
while the node was shutting down.

The test only works when the node has been restarted and is
up/processing and *then* sees the reorg.

Accomplish this by only submitting 1 transaction, rather than 2, which
should give the nodes *ample* time (36 seconds) to restart.

Previously, since we were submitting two transactions in successive
blocks, we were too close the the epoch boundary when we went to restart
the nodes.

Separately, fixes an issue with event listener management in the epoch
prune watcher.